pyproject.toml
[project]
name = "themex"
version = "0.1.0"
description = "A theme extraction tool using LLMs"
authors = [
    {name = "YX",email = "19346059+alysiayx@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "spacy (>=3.8.4,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "openai (>=1.68.2,<2.0.0)",
    "transformers @ git+https://github.com/shumingma/transformers.git",
    "llama-cpp-python (>=0.3.8,<0.4.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "tf-keras (>=2.19.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "anthropic (>=0.49.0,<0.50.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "accelerate (>=1.5.2,<2.0.0)",
    "json-repair (>=0.41.1,<0.42.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "jsonpointer (>=3.0.0,<4.0.0)",
    "langchain-core (>=0.3.54,<0.4.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "jsonfinder (>=0.4.2,<0.5.0)",
    "azure-ai-textanalytics (>=5.3.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"